handle export errors and surface to the user
